feat: React NodeJS Refactorización
- Añado "components/Items.tsx" y muevo el proceso de dibujado de la lista
- Añado "hooks/useItems.ts" y muevo la lógica en un hook propio.
- Arreglo el tipo del id en "models/item.model.tes".

#20231114e
