+++ 20231114_ReactNodeJS/src/models/items.model.ts
@@ -1,5 +1,6 @@
+export type ItemId = `${string}-${string}-${string}-${string}-${string}`;
 export interface ItemsInterface{
-  id:`${string}-${string}-${string}-${string}-${string}`, /*string,*/
+  id:ItemId, /*string,*/
   timestamp: number,
   text: string,
 }
